added additional vhosts
The WebOps team decided that we would like a way to differentiate
between HTTP and HTTPS requets. In order to address this, for each
redirect I've added two vhosts. If the request protocol was HTTPS,
it will write the request to the HTTPS log file - same with HTTP.

This is just an experiment while I look at options to address this
need.